wip(page): add unit test
